requirements.txt
crewai>=0.36.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
PyGithub>=2.3.0
requests>=2.31.0
